+++ src/na/pages/Blog.tsx
@@ -134,7 +134,7 @@ function BlogArticle() {
 
   return (
     <>
-      <PageMeta title={post.title} description={post.excerpt || ''} image={post.thumbnail} url={`/en/blog/${post.slug}`} />
+      <PageMeta title={post.title} description={post.excerpt || ''} image={post.thumbnail} url={`/en/blog/${post.slug}`} article={{ date: post.date, author: post.author }} />
       {/* Hero */}
       <section style={{
         backgroundColor: 'var(--matera-blue)',