+++ src/na/pages/Press.tsx
@@ -141,7 +141,7 @@ function PressArticle() {
 
   return (
     <>
-      <PageMeta title={post.title} description={post.excerpt || ''} image={post.thumbnail} url={'/en/press/' + post.slug} />
+      <PageMeta title={post.title} description={post.excerpt || ''} image={post.thumbnail} url={'/en/press/' + post.slug} article={{ date: post.date }} />
       {/* Hero */}
       <section style={{
         backgroundColor: 'var(--matera-blue)',